refactor: changed routes
