+++ src/app/professor/select-area/page.tsx
@@ -1,3 +1,4 @@
+import { Header } from "@/app/_components/header";
 import { getAreas } from "./actions";
 import { AreaForm } from "./area-form";
 
@@ -6,6 +7,7 @@ const SelectArea = async () => {
 
   return (
     <section className="flex flex-col h-screen">
+      <Header></Header>
       <AreaForm data={data} />
     </section>
   );
